// apps/enterprise-service-management/src/generated/models/automation-condition-model.ts
import type { AutomationRule } from './automation-rule-model';
import type { Workspace } from './workspace-model';

export const AutomationConditionGroupLogicKeyToLabel = {
  'AND': 'AND',
  'OR': 'OR'
} as const;
export type AutomationConditionGroupLogicKey = keyof typeof AutomationConditionGroupLogicKeyToLabel;

export const AutomationConditionOperatorKeyToLabel = {
  'Equals': 'Equals',
  'NotEquals': 'Not Equals',
  'GreaterThan': 'Greater Than',
  'GreaterThanOrEquals': 'Greater Than or Equals',
  'LessThan': 'Less Than',
  'LessThanOrEquals': 'Less Than or Equals',
  'Contains': 'Contains',
  'IsEmpty': 'Is Empty',
  'IsNotEmpty': 'Is Not Empty'
} as const;
export type AutomationConditionOperatorKey = keyof typeof AutomationConditionOperatorKeyToLabel;

export interface AutomationCondition {
  /**
   * @displayName Automation Condition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Condition Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  conditionName: string;
  /**
   * @displayName Active
   * @validationRule Required for create/update operations
   */
  active: boolean;
  /**
   * @displayName Automation Rule
   * @validationRule Required for create/update operations
   */
  automationRule: Pick<AutomationRule, 'id' | 'name1'>;
  /**
   * @displayName Comparison Value
   * @validationRule Required for create/update operations
   */
  comparisonValue: string;
  /**
   * @displayName Condition Group
   * @validationRule Required for create/update operations
   */
  conditionGroup: number;
  /**
   * @displayName Condition Order
   * @validationRule Required for create/update operations
   */
  conditionOrder: number;
  /**
   * @displayName Field Name
   * @validationRule Required for create/update operations
   */
  fieldName: string;
  /**
   * @displayName Group Logic
   * @validationRule Required for create/update operations
   */
  groupLogicKey: AutomationConditionGroupLogicKey;
  /**
   * @displayName Operator
   * @validationRule Required for create/update operations
   */
  operatorKey: AutomationConditionOperatorKey;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _AutomationCondition = 'AutomationCondition' as const;