// apps/enterprise-service-management/src/generated/models/automation-rule-model.ts
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const AutomationRuleStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type AutomationRuleStatusKey = keyof typeof AutomationRuleStatusKeyToLabel;

export interface AutomationRule {
  /**
   * @displayName Automation Rule ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Catalog Item Code
   */
  catalogItemCode?: string;
  /**
   * @displayName Cooldown Minutes
   * @validationRule Required for create/update operations
   */
  cooldownMinutes: number;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Destination Status
   */
  destinationStatus?: string;
  /**
   * @displayName Maximum Chain Depth
   * @validationRule Required for create/update operations
   */
  maximumChainDepth: number;
  /**
   * @displayName Maximum Executions Per Request
   * @validationRule Required for create/update operations
   */
  maximumExecutionsPerRequest: number;
  /**
   * @displayName Rule Order
   * @validationRule Required for create/update operations
   */
  ruleOrder: number;
  /**
   * @displayName Run As Person
   * @validationRule Required for create/update operations
   */
  runAsPerson: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Service Code
   */
  serviceCode?: string;
  /**
   * @displayName Source Status
   */
  sourceStatus?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: AutomationRuleStatusKey;
  /**
   * @displayName Stop Processing
   * @validationRule Required for create/update operations
   */
  stopProcessing: boolean;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _AutomationRule = 'AutomationRule' as const;